+++ src/pages/CastMember/CastMember.js
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 
 import { PersonDetails } from "./CastMember.styled";
 
-import { dateHelper } from "../../../../../utils/helperFunctions";
+import { dateHelper } from "../../utils/helperFunctions";
 
 const CastMember = (props) => {
   const { id } = useParams();
