+++ src/pages/Home/store/actions/homeActionTypes.js
@@ -1,4 +1,7 @@
 export const GET_ALL_TV_SHOWS_REQUEST = "GET_ALL_TV_SHOWS_REQUEST";
 export const GET_ALL_TV_SHOWS_SUCCESS = "GET_ALL_TV_SHOWS_SUCCESS";
 export const GET_ALL_TV_SHOWS_FAILURE = "GET_ALL_TV_SHOWS_FAILURE";
+export const SEARCH_TV_SHOWS_REQUEST = "SEARCH_TV_SHOWS_REQUEST";
+export const SEARCH_TV_SHOWS_SUCCESS = "SEARCH_TV_SHOWS_SUCCESS";
+export const SEARCH_TV_SHOWS_FAILURE = "SEARCH_TV_SHOWS_FAILURE";
 export const SET_DROPDOWN_VALUE = "SET_DROPDOWN_VALUE";
