+++ src/pages/SeasonsList/SeasonsList.styled.js
