+++ src/pages/ShowDetails/store/actions/showDetailsActionTypes.js
