--- src/store/actions/searchActionTypes.js
@@ -1,3 +1,0 @@
-export const SEARCH_TV_SHOWS_REQUEST = "SEARCH_TV_SHOWS_REQUEST";
-export const SEARCH_TV_SHOWS_SUCCESS = "SEARCH_TV_SHOWS_SUCCESS";
-export const SEARCH_TV_SHOWS_FAILURE = "SEARCH_TV_SHOWS_FAILURE";
